Updated to Julia 0.3.0-prerelease
Updated MPI wrappers to work with the latest julia source.
A number of minor changes.
I had to comment out the finalizer() call in line 34 of mpi-base.jl to
get OpenMPI to exit gracefully. I am guessing the garbage collection is
getting in the way. Any suggestions here would be appreciated. Also the
testing framework has changed and so a number of changes were
necessary.
I am a complete novice with julia, however, so I hope all the changes
are sensible. At least all the tests to now pass for me.