Move to C API
Works around the issues encountered last time (#169) by generating a small C
program which in turn generates the necessary constants, types and
offsets. To avoid the problem of link-time definitions, it uses the
Fortran constants for all handles, which are converted to the C variants
at `MPI_Init`.

This approach has several benefits:
 - No need for CMake or a Fortran compiler
 - No implicit use of gfortran ABI for strings
 - Simpler to extend, as the new functions only need to be added in one place.

I've also taken the liberty splitting the source into several files
based on functionality, which should make it easier to find things.