Throw an error when using custom reductions on non-Intel platforms (#473)
The error thrown by Julia when getting to closures unsupported on non-Intel
platforms is like this:

```
ERROR: LoadError: cfunction: closures are not supported on this platform
in expression starting at /path/to/file.jl:line
```

and nothing else, without a full stacktrace, but `/path/to/file.jl:line` is only
the top-level frame of the stack.  It can take a while to understand where the
error is actually coming from especially if several packages are involved,
provided that you're willing to debug it, otherwise users are out of luck.

This change makes the origin of the error clearer.