Base `MPI.identify_implementation` on helper function in `MPIPreferences`
We basically moved the detection of ABI/implementation to `MPIPreferences`, and
added only there the support for new ABIs/implementations, which means that
`MPI.identify_implementation` is now giving incorrect results for them, and also
the constants `MPI.MPI_LIBRARY` and `MPI.MPI_LIBRARY_VERSION` are now wrong.

By having a new helper function in `MPIPreferences` we can have all other
functions use it under the hood and keep everything in-sync.